+++ CHANGELOG.md
@@ -4,10 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-[Unreleased]
+## [Unreleased]
 ### Added
 - V3 api with support for paging
 - Additional debugging options including debug_on_failure
+- ARA links in the component and session records
 
 ### Changed
 - Disabled concurrent Jenkins builds on same branch/commit